+++ R/TaskCompRisks.R
@@ -13,7 +13,12 @@
 #'
 #' The `task_type` is set to `"cmprsk"`.
 #'
-#' **Note:** Currently only right-censoring is supported.
+#' @details
+#' The following design choices are made for this task:
+#' - Currently only right-censoring is supported.
+#' - Competing risks tasks must always contain **at least two non-censoring event types**, i.e \eqn{K \geq 2}.
+#' Use stratified resampling (for example via the `"stratum"` role) to reduce the risk
+#' of creating training splits with fewer than two causes.
 #'
 #' @template param_rows
 #'
@@ -194,20 +199,34 @@ TaskCompRisks = R6Class(
     #' @description
     #' Subsets the task, keeping only the rows specified via row ids `rows`.
     #' This operation mutates the task in-place.
+    #' A warning is thrown if the filtering results in fewer competing events
+    #' than the original task.
+    #' An error is thrown if fewer than two competing events remain after filtering.
     #'
     #' @return Returns the object itself, but modified **by reference.**
     filter = function(rows = NULL) {
-      # check that we don't remove the competing events from the data
-      uevents = self$unique_events(rows)
-      if (length(uevents) != length(self$cmp_events)) {
-        stopf(
-          "Can't filter task %s: %i competing events found, but row filtering results in %i unique competing event(s)",
+      n_events_before = length(self$cmp_events)
+      event_levels_after = levels(as.factor(self$event(rows)))
+      n_events_after = length(setdiff(event_levels_after, "0"))
+
+      if (n_events_after < 2L) {
+        error_input(
+          "Can't filter task %s: row filtering leaves %i competing event(s), but at least 2 are required",
           self$id,
-          length(self$cmp_events),
-          length(uevents)
+          n_events_after
         )
       }
 
+      if (n_events_after < n_events_before) {
+        warning_mlr3(
+          "While filtering task %s: %i competing events found, but row filtering results in %i unique competing events",
+          self$id,
+          n_events_before,
+          n_events_after
+        )
+      }
+
+      private$.event_levels = event_levels_after
       super$filter(rows)
     }
   ),